# serveur-REST-depenses-v0.2/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Serveur des dépenses partagées",
        "description": "Utilisation d'un serveur REST",
        "contact": {
            "email": "fred.hemery@univ-artois.fr"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Le serveur pour les dépenses partagées"
        }
    ],
    "paths": {
...
        "/depenses": {
            "get": {
                "tags": [
                    "Depenses"
                ],
                "summary": "Renvoie les informations de toutes les dépenses",
                "description": "Renvoie les dépenses",
                "operationId": "show",
                "responses": {
                    "200": {
                        "description": "dépense response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "format": "int64"
                                                            },
                                                            "description": {
                                                                "type": "string"
                                                            },
                                                            "date_depense": {
                                                                "type": "string"
                                                            },
                                                            "montant": {
                                                                "type": "number",
                                                                "format": "float"
                                                            },
                                                            "acheteur": {
                                                                "type": "object",
                                                                "allOf": [
                                                                    {
                                                                        "properties": {
                                                                            "id": {
                                                                                "type": "integer",
                                                                                "format": "int64"
                                                                            },
                                                                            "email": {
                                                                                "type": "string"
                                                                            },
                                                                            "nom": {
                                                                                "type": "string"
                                                                            },
                                                                            "prenom": {
                                                                                "type": "string"
                                                                            },
                                                                            "avatar": {
                                                                                "type": "string"
                                                                            }
                                                                        },
                                                                        "type": "object"
                                                                    }
                                                                ]
                                                            },
                                                            "participants": {
                                                                "type": "array",
                                                                "items": {
                                                                    "properties": {
                                                                        "id": {
                                                                            "type": "integer",
                                                                            "format": "int64"
                                                                        },
                                                                        "nom": {
                                                                            "type": "string"
                                                                        },
                                                                        "prenom": {
                                                                            "type": "string"
                                                                        },
                                                                        "avatar": {
                                                                            "type": "string"
                                                                        },
                                                                        "created_at": {
                                                                            "type": "string"
                                                                        },
                                                                        "updated_at": {
                                                                            "type": "string"
                                                                        },
                                                                        "pivot": {
                                                                            "type": "object",
                                                                            "allOf": [
                                                                                {
                                                                                    "properties": {
                                                                                        "depense_id": {
                                                                                            "type": "integer",
                                                                                            "format": "int64"
                                                                                        },
                                                                                        "membre_id": {
                                                                                            "type": "integer",
                                                                                            "format": "int64"
                                                                                        },
                                                                                        "quote_part": {
                                                                                            "type": "integer",
                                                                                            "format": "float"
                                                                                        }
                                                                                    },
                                                                                    "type": "object"
                                                                                }
                                                                            ]
                                                                        }
                                                                    },
                                                                    "type": "object"
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:membres",
                            "read:membres"
                        ]
                    }
                ]
            }
        },
        "/depenses/{id}": {
            "get": {
                "tags": [
                    "Depenses"
                ],
                "summary": "Renvoie les informations d'une dépense",
                "description": "Renvoie les informations d'une dépense",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id d'une dépense",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "response: les données d'une dépense",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "format": "int64"
                                                },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "date_depense": {
                                                    "type": "string"
                                                },
                                                "montant": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "acheteur": {
                                                    "type": "object",
                                                    "allOf": [
                                                        {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "format": "int64"
                                                                },
                                                                "email": {
                                                                    "type": "string"
                                                                },
                                                                "nom": {
                                                                    "type": "string"
                                                                },
                                                                "prenom": {
                                                                    "type": "string"
                                                                },
                                                                "avatar": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    ]
                                                },
                                                "participants": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "format": "int64"
                                                            },
                                                            "nom": {
                                                                "type": "string"
                                                            },
                                                            "prenom": {
                                                                "type": "string"
                                                            },
                                                            "avatar": {
                                                                "type": "string"
                                                            },
                                                            "created_at": {
                                                                "type": "string"
                                                            },
                                                            "updated_at": {
                                                                "type": "string"
                                                            },
                                                            "pivot": {
                                                                "type": "object",
                                                                "allOf": [
                                                                    {
                                                                        "properties": {
                                                                            "depense_id": {
                                                                                "type": "integer",
                                                                                "format": "int64"
                                                                            },
                                                                            "membre_id": {
                                                                                "type": "integer",
                                                                                "format": "int64"
                                                                            },
                                                                            "quote_part": {
                                                                                "type": "integer",
                                                                                "format": "float"
                                                                            }
                                                                        },
                                                                        "type": "object"
                                                                    }
                                                                ]
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:membres",
                            "read:membres"
                        ]
                    }
                ]
            }
        },
...
    },
...
}